Add support for listing WiFi networks visible to device
The supported smart devices can list visible networks. This is
presumably useful when trying to confirm a network is visible from
where the plug is placed.

Debugging this, I noticed that there is a 1028 byte limit on
single reply chunks, so I now loop data reply reads. One of my
devices generates more than 1028 bytes of reply to a WiFi AP listing.

Improved the latency for --scan. Previously, it was taking a long time
to timeout on some network addresses.

Signed-off-by: Tinkerer <tinkerer@zappem.net>